Make token canister impl use stable state instead of state field
This is because we now have `AuctionState` in place that is also
a stable state that needs to be upgraded. Since we now have two
stable state fields (one for token state and one for auction state)
we need to upgrade the correspondingly. To avoid further confusion
I decided to move state out of `CanisterState` and use crate-local
`CanisterState::get()` state instead.